+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changes in 3.3.0
 
+- upgrade django-cas library
 - removed SSL setting from MDID; the application should now
   be protected by enforcing SSL for the whole site on the
   web server level
 - added files for Docker support
+- additional documentation